+++ internal/pkg/usecase/buildroute/config.go
@@ -32,7 +32,7 @@ type (
 
 	FaultyPoolsConfig struct {
 		// Min slippage threshold configured in BPS format, ex: 0.01% -> 1, 0.5% -> 50
-		MinSlippageThreshold int64           `mapstructure:"minSlippageThreshold" json:"minSlippageThreshold"`
+		MinSlippageThreshold float64         `mapstructure:"minSlippageThreshold" json:"minSlippageThreshold"`
 		WhitelistedTokenSet  map[string]bool `mapstructure:"whitelistedTokenSet" json:"whitelistedTokenSet"`
 	}
 )
